Release of today's hacking
This is the second last day of hacking and the code is gradually coming
together. We still have some small bugs and so on before the first
formal beta is ready, but you can play with this Ruby file, this XML
input and this SQL output to get the idea of where we are going.