Released version 0.13
This adds progress feedback at the command line which improves on our
previous one item = one "." display. That got old pretty quickly when
processing big files.

The new progress display shows a count of items read. It's not perfect,
as we are overwritingn the same line many times, and occasionally you
see the old numbers as well as the new, but it's an improvement.